+++ src/modules/expansion/expansion.service.dbbackup.ts
@@ -14,22 +14,25 @@ import { Injectable } from '@nestjs/common';
 import { CloudStorageService } from '@app/processors/helper/helper.service.cs';
 import * as APP_CONFIG from '@app/app.config';
 
-// Configs
-const UPFAILE_TIMEOUT = 1000 * 60 * 5;
+// Config
+const UP_FAILED_TIMEOUT = 1000 * 60 * 5;
 const UPLOAD_INTERVAL = '0 0 3 * * *';
+const BACKUP_FILE_NAME = 'nodepress.tar.gz'
+const BACKUP_DIR_PATH = path.join(APP_CONFIG.APP.PROJECT_ROOT_PATH, 'dbbackup')
+const BACKUP_DATA_PATH = path.join(BACKUP_DIR_PATH, BACKUP_FILE_NAME)
 
-const BACKUP_FILE_EXT = '.tar.gz';
-const BACKUP_SHELL_PATH = path.normalize(APP_CONFIG.DB_BACKUP.backupShellPath);
-const BACKUP_DATA_PATH = path.resolve(APP_CONFIG.DB_BACKUP.backupFilePath, `nodepress${BACKUP_FILE_EXT}`);
+const SHELL_DIR_PATH = path.join(APP_CONFIG.APP.PROJECT_ROOT_PATH, 'scripts')
+const BACKUP_SHELL_PATH = path.normalize(path.join(SHELL_DIR_PATH, 'dbbackup.sh'));
+// const RECOVER_SHELL_PATH = path.normalize(path.join(SHELL_DIR_PATH, 'dbrecover.sh'));
 
 @Injectable()
-export class DBBackupcService {
+export class DBBackupService {
 
   constructor(private readonly cloudStorageService: CloudStorageService) {
     console.log('DB Backup 开始执行定时数据备份任务！');
     schedule.scheduleJob(UPLOAD_INTERVAL, () => {
       this.backup().catch(() => {
-        setTimeout(this.backup, UPFAILE_TIMEOUT);
+        setTimeout(this.backup, UP_FAILED_TIMEOUT);
       });
     });
   }
@@ -43,7 +46,7 @@ export class DBBackupcService {
 
       shell.exec(`sh ${BACKUP_SHELL_PATH}`, (code, out) => {
         const fileDate = moment(new Date()).format('YYYY-MM-DD-HH:mm');
-        const fileName = `nodepress-db-backup-${fileDate}${BACKUP_FILE_EXT}`;
+        const fileName = `nodepress-db-backup-${fileDate}.tar.gz`;
         console.info('DB Backup shell 执行完成！', code, out);
         console.info('DB Backup 上传文件: ' + fileName);
         console.info('DB Backup 文件源位置: ' + BACKUP_DATA_PATH);