+++ src/transformers/urlmap.transformer.ts
@@ -23,14 +23,10 @@ export function getGuestbookPageUrl(): string {
   return `${APP_CONFIG.APP.URL}/guestbook`;
 }
 
-export function getLensPageUrl(): string {
-  return `${APP_CONFIG.APP.URL}/lens`;
-}
-
 export function getAboutPageUrl(): string {
   return `${APP_CONFIG.APP.URL}/about`;
 }
 
-export function getSitemapPageUrl(): string {
-  return `${APP_CONFIG.APP.URL}/sitemap`;
+export function getArchivePageUrl(): string {
+  return `${APP_CONFIG.APP.URL}/archive`;
 }